sar/sadf: Use size_t type for size variable in read() syscall
Use size_t type expected by read() system call to make sure size
variable cannot be negative or overflow integer capacity.

+++ sa.h
@@ -1289,7 +1289,7 @@ void remap_struct
 void replace_nonprintable_char
 	(int, char *);
 int sa_fread
-	(int, void *, int, int);
+	(int, void *, size_t, int);
 int sa_get_record_timestamp_struct
 	(unsigned int, struct record_header *, struct tm *, struct tm *);
 int sa_open_read_magic
